refactor: organizes TextStyle exports in their respective folder
