+++ src/components/ui/text-style/index.ts
@@ -1,0 +1,2 @@
+export { default as TextStyle } from './TextStyle.vue';
+export type { TextStyleProps, TextStyleVariant } from './types';
